// src/components/Header/types.d.ts
export interface HeaderProps{
  /**
   * @description URL for Avatar Image
   */
  avatarUri?: string;
  /**
   * @description URL For Brand Image
   */
  brandUri?: string;

}

export interface ColProps {
  /**
   * @description align-content acepts all flex align values
   */
  align?: string;
  /**
   * @description justify content acepts all flex align values
   */
  justify?: string;
}

export interface AvatarProps {
  size?: number;
}